Move effects into test-report
